API Endpoint for next Cell in ProductionLine with automatic OrdinalNumber

+++ CellTracker/CellTracker.Api/Models/Cell.cs
@@ -11,7 +11,7 @@ namespace CellTracker.Api.Models
         [MaxLength(35)]
         public string Name { get; set; }
 
-        public short OrdinalNumber { get; set; }
+        public int OrdinalNumber { get; set; }
 
         [MaxLength(200)]
         public string? Description { get; set; }  // optional Description,
